requirements.txt
# SentinelWeb - Dependências do Projeto
# Framework Web Assíncrono
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Banco de Dados
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Templates e Formulários
jinja2==3.1.3
python-multipart==0.0.6

# Autenticação
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# HTTP Requests (para verificação de uptime)
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0  # Para envio de alertas Telegram

# Celery e Redis (Workers em Background)
celery[redis]==5.3.4
redis==4.6.0

# Utilitários
python-dotenv==1.0.0
pyOpenSSL==24.0.0

# Validação de dados
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Whois (verificação de expiração de domínio)
python-whois==0.8.0
python-dateutil==2.8.2

# DNS Queries para Blacklist (RBL)
dnspython==2.4.2

# Visual Regression Testing
playwright==1.40.0
Pillow==10.2.0
numpy==1.26.3

# Tech Stack Detection & Security
python-Wappalyzer==0.3.1
packaging==23.2

